requirements.txt
asgiref==3.9.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cffi==2.0.0
channels==4.1.0
chardet==5.2.0
constantly==23.10.4
crispy-tailwind==0.5.0
cryptography==46.0.3
daphne==4.1.2
Django==5.2.5
django-crispy-forms==2.1
django-extensions==3.2.3
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
openpyxl==3.1.5
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-decouple==3.8
redis>=5.0.0
reportlab==4.2.2
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.6.0
zope.interface==8.1.1
